Extract custom model builder helpers
